deckGUI/Gui/requirements.txt
pillow
webbrowser